Added `=>>` operator to `->` macro and `<<=` to `->>`
Added opertors to the existing `->` and `->>` threading macros to
overwrite the position of the threaded value. This is useful in Scheme
because, annoyingly, many built-in Scheme/Racket functions that operate
on lists take the list argument either as the first or last argument.
For example, `sort` takes the list as the first argument, but `map`
and `fold(l|r)` take the list as the last argument.

Including `=>>` as the first element in a function call's form in `->`
passes the threaded value as the last argument instead of the first,
while including `<<=` in a similar fashion within `->>` passes the
threaded argument as the first argument. For example, the following
are all equivalent, producing '(3 6 9)

  (-> '(5 8 2)
    (sort <)
    (=>> map (lambda (x) (+ x 1))))

  (->> '(5 8 2)
    (<<= sort <)
    (map (lambda (x) (+ x 1))))

  (any-> '(5 8 2) lst
    (sort lst <)
    (map (lambda (x) (+ x 1)) lst))